requirements.txt
pytest==6.2.5
requests==2.26.0
pytest-selenium==2.0.1
assertpy==1.1